requirements.txt
Pillow==6.1.0
django-ckeditor==5.8.1
pylint==2.3.1
mysql==0.0.2
mysqlclient==1.4.2.post1